Pytorch-UNet-master/requirements.txt
matplotlib==3.6.2
numpy==1.23.5
Pillow==9.3.0
tqdm==4.64.1
TensorBoardX
opencv-python